add verification and error messages

+++ src/step-2-install-gpu-driver.sh
@@ -15,14 +15,21 @@ DRIVER_PACKAGE=APM_470.10.10_5.11.0-1028.31.tar
 
 ## Install gpu driver required dependency and driver itself. It will reboot the system at the end.
 install_gpu_driver(){
-    echo "Installing driver"
-    sudo apt-get install -y libgl1 binutils xserver-xorg-core
+    current_kernel=$(uname -r)
+    if [ "$current_kernel" != "$REQUIRED_KERNEL_VERSION" ]; 
+    then
+    	echo "Current kernel version: ($current_kernel), expected: ($REQUIRED_KERNEL_VERSION)."
+    	echo "Please retry step-1-install-kernel."
+    else 
+    	echo "kernel verified successfully, start driver installation."
+	    sudo apt-get install -y libgl1 binutils xserver-xorg-core
 
-    sudo systemctl set-default multi-user.target
-    mkdir apm470driver
-    tar -xvf $DRIVER_PACKAGE --directory apm470driver
-    sudo dpkg -i apm470driver/*.deb
-    sudo reboot
+	    sudo systemctl set-default multi-user.target
+	    mkdir apm470driver
+	    tar -xvf $DRIVER_PACKAGE --directory apm470driver
+	    sudo dpkg -i apm470driver/*.deb
+	    sudo reboot
+	fi
 }
 
 if [[ "${#BASH_SOURCE[@]}" -eq 1 ]]; then
